% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alley_2000}
\alias{alley_2000}
\title{Richard Alley’s (2000) Greenland ice core data.}
\format{
An object of class \code{tibble}.
}
\source{
\url{https://www.ncei.noaa.gov/pub/data/paleo/icecore/greenland/summit/gisp2/isotopes/gisp2_temp_accum_alley2000.txt}
}
\usage{
alley_2000
}
\description{
A \code{tibble} containing Richard Alley’s (2000) Greenland ice core data.
}
\keyword{datasets}
