% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bacon_age_models}
\alias{bacon_age_models}
\title{Bacon Age-depth Models.}
\format{
An object of class \code{tibble}.
}
\source{
\url{https://www.neotomadb.org/}
}
\usage{
bacon_age_models
}
\description{
A \code{tibble} containing  Bacon age models for fossil pollen sites.
}
\keyword{datasets}
