% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPtoBCAD.R
\name{BPtoBCAD}
\alias{BPtoBCAD}
\title{Convert BP dates to BC/AD format}
\usage{
BPtoBCAD(x)
}
\arguments{
\item{x}{A numerical vector.}
}
\value{
A vector with BC/BCE dates expressed as negative numbers and AD/CE dates as positive ones.
}
\description{
Converts calibrated BP dates to BC/AD dates
}
\examples{
BPtoBCAD(5500)
}
