% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gdd.R
\name{calc_gdd}
\alias{calc_gdd}
\title{Calculate growing degree days}
\usage{
calc_gdd(tmin, tmax, t.base, t.cap = NULL)
}
\arguments{
\item{tmin}{A numeric vector.}

\item{tmax}{A numeric vector.}

\item{t.base}{A numeric vector.}

\item{t.cap}{A numeric vector.}
}
\value{
A vector with growing degree days.
}
\description{
Calculate growing degree days from PRISM temperature data
}
