% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_pollen.R
\name{compile_pollen}
\alias{compile_pollen}
\title{Compile Modern Pollen Data}
\usage{
compile_pollen(x)
}
\arguments{
\item{x}{A download_list from Neotoma.}
}
\value{
A tibble with modern pollen data.
}
\description{
Compile modern pollen data into a \code{tibble}
}
