% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coreTops}
\alias{coreTops}
\title{Core tops dataset extracted from the fossil pollen dataset.}
\format{
An object of class \code{tibble}.
}
\source{
\url{https://www.neotomadb.org/}
}
\usage{
coreTops
}
\description{
A \code{tibble} containing the core tops samples for United States, Canada, and Mexico downloaded from Tilia in Neotoma.
}
\keyword{datasets}
