% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_prism_normals.R
\name{extract_prism_normals}
\alias{extract_prism_normals}
\title{Extract PRISM Climate Data at Site Locations}
\usage{
extract_prism_normals(sites)
}
\arguments{
\item{sites}{A numeric vector.}
}
\value{
A vector with growing degree days.
}
\description{
Extract PRISM Climate Data for July for 1961 to 1990 at Site Locations
}
