% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fossil_pollen}
\alias{fossil_pollen}
\title{Fossil Pollen dataset for SWUS across all time.}
\format{
An object of class \code{tibble}.
}
\source{
\url{https://www.neotomadb.org/}
}
\usage{
fossil_pollen
}
\description{
A \code{tibble} containing proportional fossil pollen for the SWUS.
}
\keyword{datasets}
