% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fossil_west_post5500}
\alias{fossil_west_post5500}
\title{Fossil pollen dataset for SWUS for post 5500 BP.}
\format{
An object of class \code{tibble}.
}
\source{
\url{https://www.neotomadb.org/}
}
\usage{
fossil_west_post5500
}
\description{
A \code{tibble} containing the fossil pollen dataset for SWUS for post 5500 BP and that have ages for each sample.
}
\keyword{datasets}
