% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coreTops.R
\name{get_coreTops}
\alias{get_coreTops}
\title{Get Core Top Data from Fossil Pollen Cores}
\usage{
get_coreTops()
}
\value{
A tibble with dataset ID, site ID, and sample ID.
}
\description{
Get pollen core top sample numbers from fossil pollen cores
}
