% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fossil_pollen.R
\name{get_fossil_pollen}
\alias{get_fossil_pollen}
\title{Download Fossil Pollen Data from Neotoma}
\usage{
get_fossil_pollen(...)
}
\arguments{
\item{gpid}{An integer vector}
}
\value{
A tibble with fossil pollen data.
}
\description{
Download fossil pollen data from Neotoma
}
