% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_midpoints.R
\name{get_midpoints}
\alias{get_midpoints}
\title{Get the Mid-point of a Numeric Vector}
\usage{
get_midpoints(x, dp = 2)
}
\arguments{
\item{x}{A factor.}

\item{dp}{A numeric vector.}
}
\value{
A numeric vector with mid-points.
}
\description{
Calculate the the mid-point of a numeric vector
}
