% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_modern_pollen.R
\name{get_modern_pollen}
\alias{get_modern_pollen}
\title{Download Modern Pollen Data from Neotoma}
\usage{
get_modern_pollen(...)
}
\arguments{
\item{gpid}{An integer vector}
}
\value{
A tibble with modern pollen data.
}
\description{
Download modern pollen data from Neotoma
}
