% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kaufman_2020}
\alias{kaufman_2020}
\title{Kaufman et al. (2020) terrestrial composite for 30–60 °N.}
\format{
An object of class \code{tibble}.
}
\source{
\url{https://www.ncei.noaa.gov/access/paleo-search/study/27330}
}
\usage{
kaufman_2020
}
\description{
A \code{tibble} containing Kaufman et al. (2020) terrestrial composite for 30–60 °N.
}
\keyword{datasets}
