% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moberg_2005}
\alias{moberg_2005}
\title{Moberg et al. (2005) Northern Hemisphere low-frequency component.}
\format{
An object of class \code{tibble}.
}
\source{
\url{https://www.ncei.noaa.gov/pub/data/paleo/contributions_by_author/moberg2005/nhtemp-moberg2005.txt}
}
\usage{
moberg_2005
}
\description{
A \code{tibble} containing Moberg et al. (2005) Northern Hemisphere low-frequency component.
}
\keyword{datasets}
