% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MPCT_proportions}
\alias{MPCT_proportions}
\title{Modern Pollen dataset for western United States.}
\format{
An object of class \code{tibble}.
}
\source{
\urlhttps://www.neotomadb.org/}
}
\usage{
MPCT_proportions
}
\description{
A \code{tibble} containing proportional modern pollen for western United States.
}
\keyword{datasets}
