% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NAfossil_metadata_counts}
\alias{NAfossil_metadata_counts}
\title{Fossil Pollen dataset for United States, Canada, and Mexico.}
\format{
An object of class \code{tibble}.
}
\source{
\url{https://www.neotomadb.org/}
}
\usage{
NAfossil_metadata_counts
}
\description{
A \code{tibble} containing the fossil pollen dataset for United States, Canada, and Mexico downloaded from Neotoma using the neotoma r package.
}
\keyword{datasets}
