% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{osman_2021}
\alias{osman_2021}
\title{Osman et al. (2021) globally resolved surface temperatures data.}
\format{
An object of class \code{tibble}.
}
\source{
\url{https://static-content.springer.com/esm/art\%3A10.1038\%2Fs41586-021-03984-4/MediaObjects/41586_2021_3984_MOESM3_ESM.xlsx}
}
\usage{
osman_2021
}
\description{
A \code{tibble} containing Osman et al. (2021) globally resolved surface temperatures data.
}
\keyword{datasets}
