% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pages2k_2019}
\alias{pages2k_2019}
\title{PAGES 2k Consortium (2019) temperature anomalies (50th percentile) with respect to 1961–1990 CE.}
\format{
An object of class \code{tibble}.
}
\source{
\url{https://www.ncei.noaa.gov/pub/data/paleo/pages2k/neukom2019temp/recons/Full_ensemble_median_and_95pct_range.txt}
}
\usage{
pages2k_2019
}
\description{
A \code{tibble} containing PAGES 2k Consortium (2019) temperature anomalies (50th percentile) with respect to 1961–1990 CE.
}
\keyword{datasets}
