% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prism_7}
\alias{prism_7}
\title{PRISM Climate for July Dataset}
\format{
An object of class \code{tibble}.
}
\source{
\url{https://prism.oregonstate.edu/}
}
\usage{
prism_7
}
\description{
The fossil and modern pollen datasets were originally downloaded from the
Neotoma Paleoecology Database. We processed and cleaned those data in
UUSS_MAT_Reconstruction.Rmd. The climate data was acquired from PRISM.
}
\keyword{datasets}
