% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{routson_2021}
\alias{routson_2021}
\title{Routson et al. (2021) summer temperature composite (500-year bins) for Western North America.}
\format{
An object of class \code{tibble}.
}
\source{
\url{https://doi.org/10.6084/m9.figshare.12863843.v1}
}
\usage{
routson_2021
}
\description{
A \code{tibble} containing Routson et al. (2021) summer temperature composite (500-year bins) for Western North America.
}
\keyword{datasets}
