% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_age_model.R
\name{simulate_age_model}
\alias{simulate_age_model}
\title{Simulate Age-depth Model from Range}
\usage{
simulate_age_model(data)
}
\arguments{
\item{data}{A tibble with pollen core data.}
}
\value{
A tibble with the reconstruction and new age-depth model.
}
\description{
Create an age-depth model from the minimum and maximum years.
}
