% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{states}
\alias{states}
\title{State boundaries for USA.}
\format{
An object of class \code{SpatialPolygonsDataFrame}.
}
\source{
\url{https://www.sciencebase.gov/catalog/item/4f70b219e4b058caae3f8e19}
}
\usage{
states
}
\description{
A \code{Polygons} of United States' state boundaries.
}
\keyword{datasets}
