% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temp.raster}
\alias{temp.raster}
\title{Mean July Temperature from PRISM (1961-1990).}
\format{
An object of class \code{RasterLayer}.
}
\source{
\url{http://www.prism.oregonstate.edu/}
}
\usage{
temp.raster
}
\description{
A \code{raster} containing the mean July temperature for 1961-1990 from PRISM for the southwestern US.
}
\keyword{datasets}
