% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{updated_age_models}
\alias{updated_age_models}
\title{Updated Bacon Age Models.}
\format{
An object of class \code{tibble}.
}
\source{
\url{https://www.neotomadb.org/}
}
\usage{
updated_age_models
}
\description{
A \code{tibble} containing updated Bacon age models for several fossil pollen sites.
}
\keyword{datasets}
