% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{viau_2006}
\alias{viau_2006}
\title{Viau et al. (2006) North American July temperature reconstruction.}
\format{
An object of class \code{tibble}.
}
\source{
\url{https://www.ncei.noaa.gov/pub/data/paleo/pollen/recons/northamerica/viau2006namerica-temp.xls}
}
\usage{
viau_2006
}
\description{
A \code{tibble} containing Viau et al. (2006) North American July temperature reconstruction.
}
\keyword{datasets}
