% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{west_in_states}
\alias{west_in_states}
\title{Pollen range for western part of the US.}
\format{
An object of class \code{sfc_MULTIPOLYGON}.
}
\source{
\url{https://ars.els-cdn.com/content/image/1-s2.0-S0277379108000061-mmc1.zip}
}
\usage{
west_in_states
}
\description{
A \code{MULTIPOLYGON} of western pollen ranges as defined by \href{https://doi.org/10.1016/j.quascirev.2008.01.004}{Williams and Shuman 2008}, which had eastern and western pollen splits.
}
\keyword{datasets}
